engine_fund/storage/postgres_backend.py
"""PostgreSQL storage backend for simulation results (future implementation)."""

from typing import Dict, Optional, Any
import logging

from . import StorageBackend

logger = logging.getLogger(__name__)


class PostgresStorageBackend(StorageBackend):
    """
    PostgreSQL storage backend for dashboard data.
    
    This backend stores simulation results in a PostgreSQL database using SQLAlchemy.
    It enables:
    - Shared storage across multiple workers and API instances
    - Queryable result history
    - Automatic cleanup with retention policies
    - Better scalability than file-based storage
    
    **Requirements:**
    - sqlalchemy >= 2.0.0
    - psycopg2-binary >= 2.9.0 (or asyncpg for async support)
    
    **Database Schema:**
    ```sql
    CREATE TABLE simulation_results (
        task_id VARCHAR(255) PRIMARY KEY,
        dashboard_data JSONB NOT NULL,
        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
        size_bytes INTEGER,
        INDEX idx_created_at (created_at)
    );
    ```
    
    **Usage:**
    ```python
    # Set environment variable
    export STORAGE_TYPE=postgres
    export STORAGE_POSTGRES_URL=postgresql://user:pass@localhost:5432/quant
    
    # Backend will be automatically created
    storage = get_storage_backend()
    storage.save_dashboard_data(task_id, data)
    ```
    """

    def __init__(
        self,
        connection_url: str,
        pool_size: int = 5,
        max_overflow: int = 10,
    ):
        """
        Initialize PostgreSQL storage backend.

        Args:
            connection_url: SQLAlchemy database URL (e.g., postgresql://user:pass@host/db)
            pool_size: Connection pool size
            max_overflow: Maximum overflow connections

        Raises:
            ImportError: If SQLAlchemy or psycopg2 not installed
            Exception: If database connection fails
        """
        try:
            from sqlalchemy import create_engine, Column, String, JSON, Integer, DateTime
            from sqlalchemy.ext.declarative import declarative_base
            from sqlalchemy.orm import sessionmaker
            from datetime import datetime
        except ImportError as e:
            raise ImportError(
                "PostgreSQL backend requires SQLAlchemy and psycopg2. "
                "Install with: pip install sqlalchemy psycopg2-binary"
            ) from e

        logger.info(f"Initializing PostgresStorageBackend with URL: {connection_url}")
        
        # TODO: Implement database connection and schema
        # This is a stub for future implementation
        raise NotImplementedError(
            "PostgreSQL storage backend is not yet implemented. "
            "This is a placeholder for future cloud deployment. "
            "Use STORAGE_TYPE=file for now."
        )
        
        # Future implementation would include:
        # 1. Create SQLAlchemy engine
        # self.engine = create_engine(
        #     connection_url,
        #     pool_size=pool_size,
        #     max_overflow=max_overflow,
        #     echo=False,
        # )
        # 
        # 2. Define ORM model
        # Base = declarative_base()
        # class SimulationResult(Base):
        #     __tablename__ = "simulation_results"
        #     task_id = Column(String(255), primary_key=True)
        #     dashboard_data = Column(JSON, nullable=False)
        #     created_at = Column(DateTime, default=datetime.utcnow)
        #     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
        #     size_bytes = Column(Integer)
        #
        # 3. Create tables
        # Base.metadata.create_all(self.engine)
        #
        # 4. Create session factory
        # self.SessionLocal = sessionmaker(bind=self.engine)

    def save_dashboard_data(self, task_id: str, data: Dict[str, Any]) -> None:
        """
        Save dashboard data to PostgreSQL.

        Args:
            task_id: Unique task identifier
            data: Dashboard data dictionary

        Raises:
            NotImplementedError: This method is not yet implemented
        """
        raise NotImplementedError("PostgreSQL backend not yet implemented")
        
        # Future implementation:
        # session = self.SessionLocal()
        # try:
        #     result = SimulationResult(
        #         task_id=task_id,
        #         dashboard_data=data,
        #         size_bytes=len(json.dumps(data))
        #     )
        #     session.merge(result)  # Insert or update
        #     session.commit()
        #     logger.info(f"Saved dashboard data for task {task_id} to database")
        # except Exception as e:
        #     session.rollback()
        #     logger.error(f"Failed to save to database: {e}")
        #     raise
        # finally:
        #     session.close()

    def load_dashboard_data(self, task_id: str) -> Optional[Dict[str, Any]]:
        """
        Load dashboard data from PostgreSQL.

        Args:
            task_id: Unique task identifier

        Returns:
            Dashboard data dictionary if found, None otherwise

        Raises:
            NotImplementedError: This method is not yet implemented
        """
        raise NotImplementedError("PostgreSQL backend not yet implemented")
        
        # Future implementation:
        # session = self.SessionLocal()
        # try:
        #     result = session.query(SimulationResult).filter_by(task_id=task_id).first()
        #     if result:
        #         logger.info(f"Loaded dashboard data for task {task_id} from database")
        #         return result.dashboard_data
        #     return None
        # except Exception as e:
        #     logger.error(f"Failed to load from database: {e}")
        #     raise
        # finally:
        #     session.close()

    def delete_dashboard_data(self, task_id: str) -> bool:
        """
        Delete dashboard data from PostgreSQL.

        Args:
            task_id: Unique task identifier

        Returns:
            True if deleted, False if not found

        Raises:
            NotImplementedError: This method is not yet implemented
        """
        raise NotImplementedError("PostgreSQL backend not yet implemented")
        
        # Future implementation:
        # session = self.SessionLocal()
        # try:
        #     result = session.query(SimulationResult).filter_by(task_id=task_id).first()
        #     if result:
        #         session.delete(result)
        #         session.commit()
        #         logger.info(f"Deleted dashboard data for task {task_id} from database")
        #         return True
        #     return False
        # except Exception as e:
        #     session.rollback()
        #     logger.error(f"Failed to delete from database: {e}")
        #     raise
        # finally:
        #     session.close()

    def exists(self, task_id: str) -> bool:
        """
        Check if dashboard data exists in PostgreSQL.

        Args:
            task_id: Unique task identifier

        Returns:
            True if data exists, False otherwise

        Raises:
            NotImplementedError: This method is not yet implemented
        """
        raise NotImplementedError("PostgreSQL backend not yet implemented")
        
        # Future implementation:
        # session = self.SessionLocal()
        # try:
        #     result = session.query(SimulationResult).filter_by(task_id=task_id).first()
        #     return result is not None
        # finally:
        #     session.close()


__all__ = ["PostgresStorageBackend"]
