engine_fund/requirements.txt
# Core simulation
numpy>=1.24
scipy>=1.11
pandas>=1.3.0
matplotlib>=3.7
plotly>=5.0.0
reportlab>=4.0

# API & task queue
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dateutil>=2.8.0
